+++ src/main/java/com/otl/otl/service/TaskService.java
@@ -5,4 +5,5 @@ import com.otl.otl.domain.Task;
 import java.util.List;
 
 public interface TaskService {
+    List<Task> getAcceptedTasksByMemberEmail(String email);
 }
