+++ src/main/java/com/otl/otl/service/TaskServiceImpl.java
@@ -6,6 +6,7 @@ import com.otl.otl.domain.Task;
 
 import com.otl.otl.repository.TaskRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
+@Log4j2
 public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository taskRepository;
@@ -20,5 +22,10 @@ public class TaskServiceImpl implements TaskService {
     @Autowired
     public TaskServiceImpl(TaskRepository taskRepository) {
         this.taskRepository = taskRepository;
+    }
+
+    @Override
+    public List<Task> getAcceptedTasksByMemberEmail(String email) {
+        return taskRepository.findTaskByMemberEmailAndIsAccepted(email);
     }
 }
